apply comments

+++ source/core/mesh_controller.cc
@@ -10,11 +10,11 @@ MeshController::calculate_coarsening_factor(
   double D = 0.1;
 
   // Evaluation of the error used to control the mesh refinement.
-  double error = static_cast<signed int>(target_number_of_elements -
-                                         current_number_of_elements) /
+  double error = static_cast<double>(target_number_of_elements -
+                                     current_number_of_elements) /
                  static_cast<double>(target_number_of_elements);
-  double previous_error = static_cast<signed int>(target_number_of_elements -
-                                                  previous_number_of_elements) /
+  double previous_error = static_cast<double>(target_number_of_elements -
+                                              previous_number_of_elements) /
                           static_cast<double>(target_number_of_elements);
 
   previous_number_of_elements = current_number_of_elements;
